chores(liferay): Feedback from Liferay 7 review
