+++ frontend/sw360-portlet/src/main/java/org/eclipse/sw360/portal/portlets/homepage/MyTaskSubmissionsPortlet.java
@@ -10,7 +10,6 @@
  */
 package org.eclipse.sw360.portal.portlets.homepage;
 
-import static org.apache.log4j.Logger.getLogger;
 import static org.eclipse.sw360.portal.common.PortalConstants.MY_TASK_SUBMISSIONS_PORTLET_NAME;
 
 import java.io.IOException;
@@ -22,19 +21,12 @@ import javax.portlet.ResourceRequest;
 import javax.portlet.ResourceResponse;
 
 import com.google.common.collect.Lists;
-import com.liferay.portal.kernel.json.JSONArray;
-import com.liferay.portal.kernel.json.JSONFactoryUtil;
-import com.liferay.portal.kernel.json.JSONObject;
 
-import org.apache.log4j.Logger;
 import org.apache.thrift.TException;
-import org.eclipse.sw360.datahandler.common.ThriftEnumUtils;
-import org.eclipse.sw360.datahandler.thrift.ModerationState;
 import org.eclipse.sw360.datahandler.thrift.RequestStatus;
 import org.eclipse.sw360.datahandler.thrift.moderation.ModerationRequest;
 import org.eclipse.sw360.datahandler.thrift.users.User;
 import org.eclipse.sw360.portal.common.PortalConstants;
-import org.eclipse.sw360.portal.portlets.Sw360Portlet;
 import org.eclipse.sw360.portal.portlets.moderation.ModerationPortletUtils;
 import org.eclipse.sw360.portal.users.UserCacheHolder;
 import org.osgi.service.component.annotations.ConfigurationPolicy;
@@ -56,7 +48,7 @@ import org.osgi.service.component.annotations.ConfigurationPolicy;
     service = Portlet.class,
     configurationPolicy = ConfigurationPolicy.REQUIRE
 )
-public class MyTaskSubmissionsPortlet extends Sw360Portlet {
+public class MyTaskSubmissionsPortlet extends AbstractTasksPortlet {
     public void serveResource(ResourceRequest request, ResourceResponse response) throws IOException, PortletException {
         String action = request.getParameter(PortalConstants.ACTION);
 
@@ -82,36 +74,6 @@ public class MyTaskSubmissionsPortlet extends Sw360Portlet {
             log.error("Could not fetch your moderations from backend", e);
         }
 
-        JSONArray jsonModerations = getModerationData(moderations);
-        JSONObject jsonResult = JSONFactoryUtil.createJSONObject();
-        jsonResult.put("aaData", jsonModerations);
-
-        try {
-            writeJSON(request, response, jsonResult);
-        } catch (IOException e) {
-            log.error("Problem generating task submission list", e);
-        }
-    }
-
-    public JSONArray getModerationData(List<ModerationRequest> moderationList) {
-        JSONArray projectData = JSONFactoryUtil.createJSONArray();
-        for(ModerationRequest moderationRequest : moderationList) {
-            JSONObject jsonObject = JSONFactoryUtil.createJSONObject();
-
-            jsonObject.put("DT_RowId", moderationRequest.getId());
-            jsonObject.put("id", moderationRequest.getId());
-            jsonObject.put("name", moderationRequest.getDocumentName());
-            jsonObject.put("state", moderationState(moderationRequest.getModerationState()));
-
-            projectData.put(jsonObject);
-        }
-
-        return projectData;
-    }
-
-    private String moderationState(ModerationState moderationState) {
-        return "<span class='" + PortalConstants.TOOLTIP_CLASS__CSS + " "
-            + PortalConstants.TOOLTIP_CLASS__CSS + "-" + moderationState.getClass().getSimpleName() + "-" + moderationState.toString() + "'>"
-            + ThriftEnumUtils.enumToString(moderationState) + "</span>";
+        sendModerations(request, response, moderations);
     }
 }
