+++ frontend/sw360-portlet/src/main/resources/META-INF/resources/html/components/includes/releases/detailOverview.jspf
@@ -59,22 +59,22 @@
                         <div class="btn-toolbar" role="toolbar">
                             <div class="btn-group" role="group">
                                 <button id="changeReleaseButton" type="button" class="btn btn-primary dropdown-toggle" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
-                                    <core_rt:if test="${release.clearingState.value == 4}"><span class="badge badge-circle badge-success">&nbsp;</span></core_rt:if>
-                                    <core_rt:if test="${release.clearingState.value == 0}"><span class="badge badge-circle badge-danger">&nbsp;</span></core_rt:if>
-                                    <core_rt:if test="${release.clearingState.value == 1}"><span class="badge badge-circle badge-danger">&nbsp;</span></core_rt:if>
-                                    <core_rt:if test="${release.clearingState.value == 2}"><span class="badge badge-circle badge-warning">&nbsp;</span></core_rt:if>
-                                    <core_rt:if test="${release.clearingState.value == 3}"><span class="badge badge-circle badge-warning">&nbsp;</span></core_rt:if>
+                                    <core_rt:if test="${release.clearingState.value == 4}"><span class="badge badge-circle badge-success"></span></core_rt:if>
+                                    <core_rt:if test="${release.clearingState.value == 0}"><span class="badge badge-circle badge-danger"></span></core_rt:if>
+                                    <core_rt:if test="${release.clearingState.value == 1}"><span class="badge badge-circle badge-danger"></span></core_rt:if>
+                                    <core_rt:if test="${release.clearingState.value == 2}"><span class="badge badge-circle badge-warning"></span></core_rt:if>
+                                    <core_rt:if test="${release.clearingState.value == 3}"><span class="badge badge-circle badge-warning"></span></core_rt:if>
                                     Version ${release.version}
                                     <clay:icon symbol="caret-bottom" />
                                 </button>
                                 <div class="dropdown-menu" aria-labelledby="changeReleaseButton">
                                     <core_rt:forEach var="releaseItr" items="${component.releases}">
                                         <a class="dropdown-item" href="<portlet:renderURL ><portlet:param name="<%=PortalConstants.COMPONENT_ID%>" value="${component.id}"/><portlet:param name="<%=PortalConstants.RELEASE_ID%>" value="${releaseItr.id}"/><portlet:param name="<%=PortalConstants.PAGENAME%>" value="<%=PortalConstants.PAGENAME_RELEASE_DETAIL%>"/></portlet:renderURL>">
-                                            <core_rt:if test="${releaseItr.clearingState.value == 4}"><span class="badge badge-circle badge-success">&nbsp;</span></core_rt:if>
-                                            <core_rt:if test="${releaseItr.clearingState.value == 0}"><span class="badge badge-circle badge-danger">&nbsp;</span></core_rt:if>
-                                            <core_rt:if test="${releaseItr.clearingState.value == 1}"><span class="badge badge-circle badge-danger">&nbsp;</span></core_rt:if>
-                                            <core_rt:if test="${releaseItr.clearingState.value == 2}"><span class="badge badge-circle badge-warning">&nbsp;</span></core_rt:if>
-                                            <core_rt:if test="${releaseItr.clearingState.value == 3}"><span class="badge badge-circle badge-warning">&nbsp;</span></core_rt:if>
+                                            <core_rt:if test="${releaseItr.clearingState.value == 4}"><span class="badge badge-circle badge-success"></span></core_rt:if>
+                                            <core_rt:if test="${releaseItr.clearingState.value == 0}"><span class="badge badge-circle badge-danger"></span></core_rt:if>
+                                            <core_rt:if test="${releaseItr.clearingState.value == 1}"><span class="badge badge-circle badge-danger"></span></core_rt:if>
+                                            <core_rt:if test="${releaseItr.clearingState.value == 2}"><span class="badge badge-circle badge-warning"></span></core_rt:if>
+                                            <core_rt:if test="${releaseItr.clearingState.value == 3}"><span class="badge badge-circle badge-warning"></span></core_rt:if>
                                             Version ${releaseItr.version}
                                         </a>
                                     </core_rt:forEach>
