+++ frontend/sw360-portlet/src/main/resources/META-INF/resources/html/projects/includes/linkedProjectsEdit.jspf
@@ -30,29 +30,29 @@
 
 <div class="dialogs">
     <div id="deleteLinkedProjectDialog" class="modal fade" tabindex="-1" role="dialog">
-		<div class="modal-dialog modal-lg modal-dialog-centered modal-danger" role="document">
-		    <div class="modal-content">
-			<div class="modal-header">
-				<h5 class="modal-title">
-					<clay:icon symbol="question-circle" />
-					Delete link to project?
-				</h5>
-				<button type="button" class="close" data-dismiss="modal" aria-label="Close">
-					<span aria-hidden="true">&times;</span>
-				</button>
-			</div>
-				<div class="modal-body">
-			        <p>
+        <div class="modal-dialog modal-lg modal-dialog-centered modal-danger" role="document">
+            <div class="modal-content">
+                <div class="modal-header">
+                    <h5 class="modal-title">
+                        <clay:icon symbol="question-circle" />
+                        Delete link to project?
+                    </h5>
+                    <button type="button" class="close" data-dismiss="modal" aria-label="Close">
+                        <span aria-hidden="true">&times;</span>
+                    </button>
+                </div>
+                <div class="modal-body">
+                    <p>
                         Do you really want to remove the link to project <b data-name="project"></b>?
                     </p>
-				</div>
-			    <div class="modal-footer">
-			        <button type="button" class="btn btn-light" data-dismiss="modal">Cancel</button>
-			        <button type="button" class="btn btn-danger">Delete Link</button>
-			    </div>
-			</div>
-		</div>
-	</div>
+                </div>
+                <div class="modal-footer">
+                    <button type="button" class="btn btn-light" data-dismiss="modal">Cancel</button>
+                    <button type="button" class="btn btn-danger">Delete Link</button>
+                </div>
+            </div>
+        </div>
+    </div>
 </div>
 
 <script>
